+++ WebContent/WEB-INF/dwr.xml
@@ -15,9 +15,6 @@
     <create creator="new" javascript="TeamPrinter">
       <param name="class" value="jp.ac.oit.igakilab.dwr.multiple.TeamPrinter"/>
     </create>
-     <create creator="new" javascript="CharacterCatcher">
-      <param name="class" value="jp.ac.oit.igakilab.dwr.multiple.CharacterCatcher"/>
-    </create>
     <convert converter="bean" match="jp.ac.oit.igakilab.dwr.multiple.*" />
     <convert converter="exception" match="java.lang.Exception" />
     <convert converter="bean" match="java.lang.StackTraceElement"/>
