+++ src/jp/ac/oit/igakilab/dwr/multiple/TeamPrinter.java
@@ -9,10 +9,8 @@ public class TeamPrinter {
 		List<String> list = new ArrayList<>();
 		int userID = user.getuserID();
 		DBController reader = new DBController();
-
-		String[] result = reader.doget("SELECT slot1,slot2,slot3 FROM user WHERE userID = " + userID);
-		for (int i = 0; i < 3; ++i) {
-			list.add(reader.doget("SELECT imgaddr FROM sfchara WHERE charaID = " + result[i])[0]);
+		for (int i = 1; i <= 3; ++i) {
+			list.add(reader.doget("SELECT slot"+i+" FROM user WHERE userID = " + userID)[0]);
 		}
 
 		return list;
@@ -25,6 +23,7 @@ public class TeamPrinter {
 		int spot2 = user.getuserspot2();
 
 		DBController reader = new DBController();
+
 		String[] slotch = reader.doget("SELECT slot" + spot1 + ",slot" + spot2 + " FROM user WHERE userID = " + userID);
 
 
@@ -32,5 +31,18 @@ public class TeamPrinter {
 				+ " WHERE userID = " + userID);
 
 		return fetchslot(user);
+	}
+	public List<String> characterfetch(LoginData user) throws InvalidValueException {
+        List<String> list = new ArrayList<>();
+        int userID = user.getuserID();
+        DBController reader = new DBController();
+        int count = reader.doget("SELECT charaID FROM sfchara").length;
+        String sql;
+        for(int i=1;i<=count;i++){
+            sql = "SELECT chara" + i +" FROM user WHERE userID = " + userID;
+            	list.add(reader.doget(sql)[0]);
+        }
+
+        return list;
 	}
 }
